refactor: move mpi code into seperate file
